requirements.txt
flask
flask-WTF
FlightRadarAPI
python-dotenv
wtforms[email]